Split in 2 handler classes
